requirements.txt
# Core dependencies
openai-whisper>=20231117
google-generativeai>=0.3.0
python-dotenv>=1.0.0

# Whisper dependencies (required by openai-whisper)
torch>=2.0.0
numpy>=1.24.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
